refactor(FileDownloader): simplify FileDownloader, wrap the handler to handle affairs of queues Refs #283
